Issue #426: Skip git setup if using worktrees
If we are using a worktree we should not attempt to setup git (or hooks)
We need to skip the whole git config and devs should be informed that
git manipulation should be done outside of the container.

This patch also suggests to split the too long command we pass to
koha-shell.