Issue #454: Improve Git worktree support
This patch introduces a better way to detect worktrees inside the `ktd`
script, and adds a new environment variable with the calculated value to
simplify the logic in `run.sh`.

Both the Git dir and the worktree are added to the safe.dicrectory entry
and hooks are set after those get set.

Testing needs to be done on the hooks in the worktree use case.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>